+++ .gitignore
@@ -57,7 +57,7 @@ local_settings.py
 db.sqlite3
 
 # Flask stuff:
-instance/
+instance/env.prod.cfg
 .webassets-cache
 
 # Scrapy stuff:
