+++ Pipfile
@@ -1,0 +1,17 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+flask = "*"
+flask-restful = "*"
+sqlalchemy = "*"
+geoalchemy2 = "*"
+flask-sqlalchemy = "*"
+psycopg2 = "*"
+
+[requires]
+python_version = "3.7"
