+++ api/app/templates/project_detail.html
