+++ api/instance/env.test.cfg
@@ -1,0 +1,8 @@
+PG_USER = 'postgres'
+PG_PASS = '1234'
+PG_HOSTNAME = '127.0.0.1'
+PG_DBNAME = 'meep-test'
+
+TESTING = True
+SQLALCHEMY_TRACK_MODIFICATIONS = True
+DEBUG = True
