--- api/src/config.py
@@ -1,26 +1,0 @@
-import os
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-class Config:
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-class DevConfig(Config):
-    SQLALCHEMY_DATABASE_URI = os.environ.get('DEV_DATABASE_URL') or \
-        'sqlite:///dev.db'
-    TESTING = True
-    TEMPLATES_AUTO_RELOAD = True
-
-class TestConfig(Config):
-    SQLALCHEMY_DATABASE_URI = os.environ.get('TEST_DATABASE_URL') or \
-        'sqlite:///test.db'
-    TESTING = True
-
-class ProdConfig(Config):
-    SQLALCHEMY_DATABASE_URI = os.environ.get('PROD_DATABASE_URL')
-
-config = {
-    'dev': DevConfig,
-    'test': TestConfig,
-    'prod': ProdConfig,
-    'default': DevConfig
-}
